Make card counts public
Making card counts public allows players to change their strategy based
on the number of cards the next player has (e.g. if the next player has
only one card, you would not want to make a move that would reveal a
free spot).

This is better than the alternative solution of announcing "last card"
as failure to announce requires a penalty and players noticing the
failure. If one needs to be observant of any players failing to announce
"last card", they may as well be observing the number of cards directly.