State which situations allow and require passing
Although the situations where one must pass can be derived from the
rules stated earlier, stating these situations explicity makes them
clearer.

It is legal to pass even though there is a valid move on the Griddle
because the first choice a player makes on their turn is whether to play
from the Griddle or from their hand. If the player chooses to ignore the
Griddle and play from their hand, then the lack of a legal play in their
hand would require them to pass.

If there are no legal plays on the Griddle but there are in your hand,
it is not legal to pass because even if you choose to play from the
Griddle, the lack of legal plays on the Griddle forces you to play from
your hand (3.1.3.4).