Add preliminary rules for Servings
Serving: a four-of-a-kind on the top of a Stack.
Tray: the area in front of a player where Servings are placed.

In an attempt to reduce the chance of a deadlock situation (usually, the
four cards of a suit are distributed among two or more players and there
is nowhere to place those cards), reduce the number of dead Stacks by
allowing players to remove them, freeing up a card back or a Joker to
place another card on.

Servings are worth one point at the end for the winner, giving players
an incentive to take Servings.

The definition of a turn still needs to be revised to properly
accommodate Servings.