Restructure the turn sequence
Now that Servings can be taken at the end of the turn and it breaks the
accumulation of passes, it does not make sense to allow declaring a pass
and then take a Serving as the current rules suggest.

A turn is now split into two phases: the "cooking" phase and the
"serving" phase. The cooking phase allows the player to play a card on
to the Griddle; the serving phase allows the player to take Servings.
None of the core rules change; a Serving may still be taken in the
serving phase even when no action was made in the cooking phase.

A pass is declared if no action was taken in both the cooking phase and
the serving phase.