Add Hevea specific instructions
Hevea will be used to generate an HTML version. To maximize portability,
Hevea comments are used instead of the `hevea` package.

http://hevea.inria.fr/

Hevea apparenetly does not parse comments inside the arguments to a
command, so the copyright footer was inlined and included only once at
the end.

Hevea does not support the wrapfig environment, so Hevea is told to
ignore it.

Hevea does not support `\hrule`, so a definition is given to Hevea.

Hevea does not render inline math as expected, so add parentheses to
ensure that the meaning doesn't change.

The PDF is authoritative, so point to it in Hevea's render.